requirements.txt
# Requisitos mínimos del entorno de desarrollo
openai==1.72.0
python-dotenv==1.1.0
jupyterlab==4.4.0
ipykernel==6.29.5
PyYAML>=6.0
click==8.1.3
requests==2.31.0
